fix: Improved redirect URL handling

+++ routes/web.php
@@ -12,15 +12,15 @@ use App\Http\Controllers\WebsiteProfileController;
 Route::get('/', [HomeController::class, 'index']);
 
 Route::middleware(['auth'])->group(function () {
-    Route::get('/home', [DashboardController::class, 'index'])->name('home')->middleware('role:admin');
-    Route::resource('category', CategoryController::class)->middleware('role:admin');
-    Route::resource('catalogue', CatalogueController::class)->middleware('role:admin');
-    Route::get('/order', [OrderController::class, 'index'])->name('order.index')->middleware('role:admin');
-    Route::get('/order/create', [OrderController::class, 'create'])->name('order.create')->middleware('role:admin');
-    Route::post('/order/store', [OrderController::class,'store'])->name('order.store')->middleware('role:admin');
-    Route::get('/order/edit/{id}', [OrderController::class, 'edit'])->name('order.edit')->middleware('role:admin');
-    Route::put('/order/update/{id}', [OrderController::class, 'update'])->name('order.update')->middleware('role:admin');
-    Route::delete('/order/delete/{id}', [OrderController::class, 'destroy'])->name('order.destroy')->middleware('role:admin');
-    Route::resource('settings', WebsiteProfileController::class)->middleware('role:admin');
-    Route::resource('order-report', OrderReportController::class)->middleware('role:admin');
+    Route::get('/admin/home', [DashboardController::class, 'index'])->name('home')->middleware('role:admin');
+    Route::resource('admin/category', CategoryController::class)->middleware('role:admin');
+    Route::resource('admin/catalogue', CatalogueController::class)->middleware('role:admin');
+    Route::get('/admin/order', [OrderController::class, 'index'])->name('order.index')->middleware('role:admin');
+    Route::get('/admin/order/create', [OrderController::class, 'create'])->name('order.create')->middleware('role:admin');
+    Route::post('/admin/order/store', [OrderController::class,'store'])->name('order.store')->middleware('role:admin');
+    Route::get('/admin/order/edit/{id}', [OrderController::class, 'edit'])->name('order.edit')->middleware('role:admin');
+    Route::put('/admin/order/update/{id}', [OrderController::class, 'update'])->name('order.update')->middleware('role:admin');
+    Route::delete('/admin/order/delete/{id}', [OrderController::class, 'destroy'])->name('order.destroy')->middleware('role:admin');
+    Route::resource('admin/settings', WebsiteProfileController::class)->middleware('role:admin');
+    Route::resource('admin/order-report', OrderReportController::class)->middleware('role:admin');
 });
